Externalize Holder's find_message_to_handle, update_state
