+++ aries_vcx/tests/utils/devsetup_alice.rs
@@ -10,7 +10,7 @@ use aries_vcx::errors::error::{AriesVcxError, AriesVcxErrorKind, VcxResult};
 use aries_vcx::global::settings::DEFAULT_LINK_SECRET_ALIAS;
 use aries_vcx::handlers::connection::mediated_connection::{ConnectionState, MediatedConnection};
 use aries_vcx::handlers::issuance::holder::Holder;
-use aries_vcx::handlers::issuance::mediated_holder::get_credential_offer_messages;
+use aries_vcx::handlers::issuance::mediated_holder::{get_credential_offer_messages, holder_update_with_mediator};
 use aries_vcx::handlers::proof_presentation::prover::test_utils::get_proof_request_messages;
 use aries_vcx::handlers::proof_presentation::prover::Prover;
 use aries_vcx::handlers::proof_presentation::types::SelectedCredentials;
@@ -192,16 +192,16 @@ impl Alice {
     }
 
     pub async fn accept_credential(&mut self) {
-        self.credential
-            .update_state(
-                &self.profile.inject_anoncreds_ledger_read(),
-                &self.profile.inject_anoncreds(),
-                &self.profile.inject_wallet(),
-                &self.agency_client,
-                &self.connection,
-            )
-            .await
-            .unwrap();
+        holder_update_with_mediator(
+            &mut self.credential,
+            &self.profile.inject_anoncreds_ledger_read(),
+            &self.profile.inject_anoncreds(),
+            &self.profile.inject_wallet(),
+            &self.agency_client,
+            &self.connection,
+        )
+        .await
+        .unwrap();
         assert_eq!(HolderState::Finished, self.credential.get_state());
         assert_eq!(Status::Success.code(), self.credential.get_credential_status().unwrap());
     }
